Moved all crypto functions to cryptly. Is backwards compatible.
