--- crypto.browser.ts
@@ -1,1 +1,0 @@
-export const crypto = window.crypto || ((window as unknown) as { msCrypto: any }).msCrypto
