--- crypto.ts
@@ -1,3 +1,0 @@
-import * as WebCrypto from "node-webcrypto-ossl"
-
-export const crypto = new WebCrypto.Crypto() as Crypto
