+++ index.ts
@@ -1,16 +1,12 @@
+import { Base64, Password, TextDecoder, TextEncoder, Identifier } from "cryptly"
 import "./shim"
 import { Actor } from "./Actor"
 import { Algorithm } from "./Algorithm"
-import { Base64 } from "cryptly"
 import { Issuer } from "./Issuer"
-import { Password } from "./Password"
 import { Payload } from "./Payload"
 import { PropertyCrypto } from "./PropertyCrypto"
-import { TextDecoder } from "cryptly"
-import { TextEncoder } from "cryptly"
 import { Token } from "./Token"
 import { Verifier } from "./Verifier"
-import { Identifier } from "./Identifier"
 
 export {
 	Actor,
